ci: add GitHub workflow to build main branch
Note: since Khepri is still in development, and not available on Conan Center Index, a custom script is used to checkout & build Khepri as part of the CI build.

To facilitate this, and ensure reproducible builds, a lockfile has been created. This has the added benefit of requiring an explicit commit to bump the Khepri requirement, which will trigger CI.