ci: fix vulkan-validationlayers dependency compilation issue (#60)
The Visual C++ compiler kept failing often during builds of
`vulkan-validationlayers/1.3.224.1` with error:
> error D8040: error creating or communicating with child process

To "fix" this, we can avoid a rebuild of the
`vulkan-validationlayers/1.3.224.1` dependency in the first place.
Inspection of the logs and recipes shows that the latest
`robin-hood-hashing/3.11.5` recipe is too new for the older
`vulkan-validationlayers/1.3.224.1` dependency, so we now specify the
exact revision of `robin-hood-hashing/3.11.5` that is needed by
`vulkan-validationlayers/1.3.224.1`, which avoids a rebuild.

The build error is gone, and the overall Windows build is faster, too.