feat: add render pipelines (#72)
This change introduces render pipelines, an XML-configurable collection
of render passes that draw meshes in a certain way.

The options at the moment are limited (alpha blending, depth sorting,
etc), but this already allows transparent meshes to be properly drawn on
top of solid meshes, instead of being arbitrarily drawn before or after.
Most of the settings have been migrated from Materials.xml to the new
RenderPipelines.xml. Materials.xml now define the shaders and shader
parameters.

This concept can later be extended with post-processing steps and so on.