refactor: add const correctness to asset objects
The AssetCache should return pointer-to-const types; returned types should not be modifiable by the callers.
To make this work, the const correctness had to bubble down all the way into the renderer.

This necessitated a custom implementation of Diligent's RefCntAutoPtr, since it does not return mutable object references from a const smart pointer (like C++ does). The alternative is making all RefCntAutoPtrs in const objects `mutable`, which does not look great.