fix: incorrect material cbuffer construction
Every material's cbuffer was incorrectly constructed: the code didn't follow the proper packing rules.
It assumed every parameter starts on a 16-byte boundary, but that was wrong. In reality, every parameter is aligned to a 16-byte boundary IF adding it to the previous 16-byte block causes it to cross into a new 16-byte block. So a `float` can be packed after a `float3` in the same 'block', but a `float2` can't.

In practice, this caused a mismatched with the provided data, and the data expected by the shader.