fix: normal maps are read in sRGB space (#81)
Normals maps were, like all other textures, read in sRGB space. This is
wrong, as normal maps should not be gamma corrected when sampled: this
produces incorrect lighting results.

Attempts to identify normal maps so they could be loaded in linear space
are not feasible; they either rely on manually-maintained (and thus
error-prone) filename lists, or rely on (flaky) content detection.

Instead, the entire pipeline was switched to linear space: all textures
are read in linear space, and the output format is in linear space, too.
This also better matches the original Empire at War graphics pipeline:
it used DirectX9, which didn't support sRGB space for textures or render
targets. Shaders were expected to do the conversion, if at all.

With this change, shaders are not responsible for gamma. This requires
and explains the x2 output factor present in the original EaW shaders
that was used in placed. Without it, the image is too dark due to the
(lack of) gamma conversion.

Note: some texture-unpacking code is part of this change, as it was used
in an earlier iteration. It is now unused, but could still be useful in
the future.

Closes #71.