fix: alpha-blended materials are rendered as additive (#84)
Materials that should be alpha-blended (i.e. (SrcAlpha, InvSrcAlpha))
were instead additive-blended (i.e. (One, One)). This was caused by the
alpha-blending settings being fixed in the "Transparent" render pass,
whereas different Transparent meshes require may different alpha
blending.

Since we already create a separate graphics pipeline for each Material
(because the shaders differ), the fix was to introduce the same graphics
pipeline options that are in the render pass, in the material. Then, the
render pass options act as a default, and the material options can
override them.

Fixes #83.