feat: set initial camera
This change sets the initial camera position & orientation based on the data in XML and the map.
Orientation and zoom level is taken from XML. Position of the camera is taken from the map as the first "Player_0_Spawn_Point_Marker" game object. Note that there are typically multiple such markers, one for each faction.