ci: add missing 'gnu-config' lockfile dependency (#89)
With conan-io/conan-center-index@1f39c7d33d5c, a new dependency to
`gnu-config` was introduced into the `flex` library (used by the
non-Windows build for `bison`, which is used by `kbdcommon`, which is
used by `diligent-core`). This was done without bumping the version,
resulting in our CI builds since that date failing since the dependency
was not part of the lockfile.

Update the lockfile with the new dependency.