ci: avoid build of vulkan-validationlayers
Windows builds always rebuild `vulkan-validationlayers/1.3.224.1`. This is because of a workaround introduced in PR #60: to avoid a rebuild at that time, `robin-hood-hashing/3.11.5` was fixed to a certain revision. However, using that revision now is avoiding using prebuilt binaries from conancenter. So we can remove that 'workaround' and let the build pick it up from conancenter instead of rebuilding.